% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_function.R
\name{zonal_custom_function}
\alias{zonal_custom_function}
\title{zonal_custom_function}
\usage{
zonal_custom_function(raster, shape, custom_function)
}
\arguments{
\item{raster}{warstwa rastrowa}

\item{shape}{warstwa wektorowa}

\item{custom_function}{wlasna funkcja}
}
\description{
Funkcja zonal_custom_function() pozwala na dodanie przez uzytkownika wczesniej zdefiniowanej wlasnej funkcji i wyliczenia statystyk strefowych na jej podstawie
}
\examples{
test_custom_function <-
 function(x, na.rm=TRUE){ 
   wynik <- max(x)
   return(wynik)
 }
 
 zonal_custom_function <-
function(raster,shape,custom_function){
  extract(raster,shape,fun=custom_function)
}

}
\author{
Szymon Hetmanczyk
}
