% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurtoza.R
\name{zonal_kurtosis}
\alias{zonal_kurtosis}
\title{zonal_kurtosis}
\usage{
zonal_kurtosis(raster, shape)
}
\arguments{
\item{raster}{warstwa rastrowa}

\item{shape}{warstwa wektorowa}
}
\description{
Funkcja zonal_kurtosis() pozwala na wyliczenie wartosci kurtozy dla poszczegolnych komorek rastra
}
\examples{
zonal_kurtosis(raster,shape)

}
\author{
Szymon Hetmanczyk
}
