% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procent_mniejszy.R
\name{zonal_less_than_avg}
\alias{zonal_less_than_avg}
\title{zonal_less_than_avg}
\usage{
zonal_less_than_avg(raster, shape)
}
\arguments{
\item{raster}{warstwa rastrowa}

\item{shape}{warstwa wektorowa}
}
\description{
Funkcja zonal_less_than_avg() pozwala na wyliczenie dla poszczegolnych wartosci rastra jaki procent jest mniejszy od sredniej wartosci calego rastra
}
\examples{
zonal_less_than_avg(raster,shape)

}
\author{
Szymon Hetmanczyk
}
