% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max.R
\name{zonal_max}
\alias{zonal_max}
\title{zonal_max}
\usage{
zonal_max(raster, shape)
}
\arguments{
\item{raster}{warstwa rastrowa}

\item{shape}{warstwa wektorowa}
}
\description{
Funkcja zonal_max() pozwala na wyliczenie wartosci maksymalnej dla poszczegolnych komorek rastra
}
\examples{
zonal_max(raster,shape)

}
\author{
Szymon Hetmanczyk
}
