% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srednia.R
\name{zonal_mean}
\alias{zonal_mean}
\title{zonal_mean}
\usage{
zonal_mean(raster, shape)
}
\arguments{
\item{raster}{warstwa rastrowa}

\item{shape}{warstwa wektorowa}
}
\description{
Funkcja zonal_mean() pozwala na wyliczenie sredniej dla poszczegolnych komorek rastra
}
\examples{
zonal_mean(raster,shape)

}
\author{
Szymon Hetmanczyk
}
