% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediana.R
\name{zonal_median}
\alias{zonal_median}
\title{zonal_median}
\usage{
zonal_median(raster, shape)
}
\arguments{
\item{raster}{warstwa rastrowa}

\item{shape}{warstwa wektorowa}
}
\description{
Funkcja zonal_median() pozwala na wyliczenie wartosci mediany dla poszczegolnych komorek rastra
}
\examples{
zonal_median(raster,shape)

}
\author{
Szymon Hetmanczyk
}
