% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min.R
\name{zonal_min}
\alias{zonal_min}
\title{zonal_min}
\usage{
zonal_min(raster, shape)
}
\arguments{
\item{raster}{warstwa rastrowa}

\item{shape}{warstwa wektorowa}
}
\description{
Funkcja zonal_min() pozwala na wyliczenie wartosci minimalnej dla poszczegolnych komorek rastra
}
\examples{
zonal_min(raster,shape)

}
\author{
Szymon Hetmanczyk
}
