% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.R
\name{zonal_sd}
\alias{zonal_sd}
\title{zonal_sd}
\usage{
zonal_sd(raster, shape)
}
\arguments{
\item{raster}{warstwa rastrowa}

\item{shape}{warstwa wektorowa}
}
\description{
Funkcja zonal_sd() pozwala na wyliczenie wartosci odchylenia standardowego dla poszczegolnych komorek rastra
}
\examples{
zonal_sd(raster,shape)

}
\author{
Szymon Hetmanczyk
}
