% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skosnosc.R
\name{zonal_skewness}
\alias{zonal_skewness}
\title{zonal_skewness}
\usage{
zonal_skewness(raster, shape)
}
\arguments{
\item{raster}{warstwa rastrowa}

\item{shape}{warstwa wektorowa}
}
\description{
Funkcja zonal_skewness() pozwala na wyliczenie wartosci skosnosci dla poszczegolnych komorek rastra
}
\examples{
zonal_skewness(raster,shape)

}
\author{
Szymon Hetmanczyk
}
