% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wariancja.R
\name{zonal_var}
\alias{zonal_var}
\title{zonal_var}
\usage{
zonal_var(raster, shape)
}
\arguments{
\item{raster}{warstwa rastrowa}

\item{shape}{warstwa wektorowa}
}
\description{
Funkcja zonal_var() pozwala na wyliczenie wartosci wariancji dla poszczegolnych komorek rastra
}
\examples{
zonal_var(raster,shape)

}
\author{
Szymon Hetmanczyk
}
